+++ server/src/server.rs
@@ -12,7 +12,7 @@ use crate::{
         MatchNotification, MatchNotificationTable, NewUserNotification, NewUserNotificationAns,
     },
     user::{
-        EditUserAction, User, USER_ROLE_INACTIVE, USER_ROLE_REGULAR, USER_ROLE_SOFT_INACTIVE,
+        EditUserAction, User, USER_ROLE_INACTIVE, USER_ROLE_SOFT_INACTIVE,
         USER_ROLE_SUPERUSER,
     },
 };
@@ -52,9 +52,6 @@ const MATCH_NO_ANS: u8 = 0;
 const ACCEPT_MATCH: u8 = 1;
 #[allow(dead_code)]
 const DECLINE_MATCH: u8 = 2;
-
-pub const N_SEASON_ID: u32 = 1;
-pub const IS_SEASON_ID: u32 = 2;
 
 pub const STOP_SEASON: i64 = -1;
 pub const START_SEASON: i64 = -2;
@@ -100,126 +97,7 @@ impl DataBase
 {
     pub fn new(file: &str) -> Self
     {
-        let conn = match Connection::open(file)
-        {
-            Err(e) => panic!(format!("Could not create connection, {}", e)),
-            Ok(c) => c,
-        };
-
-        conn.execute(
-            &format!(
-                "create table if not exists users (
-                id              integer primary key autoincrement,
-                name            VARCHAR(20) not null unique,
-                elo             float  default 1500.0,
-                password_hash   varchar(64) not null,
-                uuid            varchar(36) not null,
-                user_role       smallint default {}
-                )",
-                USER_ROLE_REGULAR | USER_ROLE_SOFT_INACTIVE
-            ),
-            NO_PARAMS,
-        )
-        .expect("Creating user table");
-
-        conn.execute(
-            "create table if not exists matches (
-                id              integer primary key autoincrement,
-                epoch           bigint not null,
-                elo_diff        integer,
-                winner_elo      float,
-                loser_elo       float,
-                winner          integer,
-                loser           integer,
-                foreign key(winner) references users(id),
-                foreign key(loser) references users(id)
-                )",
-            NO_PARAMS,
-        )
-        .expect("Creating matches table");
-
-        conn.execute(
-            "create table if not exists match_notification (
-                id              integer primary key autoincrement,
-                winner_accept   smallint default 0,
-                loser_accept    smallint default 0,
-                epoch           bigint not null,
-                winner          integer,
-                loser           integer,
-                foreign key(winner) references users(id),
-                foreign key(loser) references users(id)
-                )",
-            NO_PARAMS,
-        )
-        .expect("Creating match_notification table");
-
-        conn.execute(
-            "create table if not exists new_user_notification (
-                id              integer primary key autoincrement,
-                name            VARCHAR(20) not null unique,
-                password_hash   varchar(64) not null
-            )",
-            NO_PARAMS,
-        )
-        .expect("Creating new_user_notification");
-
-
-        conn.execute(
-            "create table if not exists badges (
-                id              integer primary key autoincrement,
-                season_id       integer,
-                badge_index     integer,
-                pid             integer,
-                foreign key(pid) references users(id),
-                foreign key(season_id) references seasons(id)
-            )",
-            NO_PARAMS,
-        )
-        .expect("Creating badges table");
-
-
-        conn.execute(
-            "create table if not exists seasons (
-                id              integer primary key autoincrement,
-                start_epoch     integer
-            )",
-            NO_PARAMS,
-        )
-        .expect("Creating season table");
-
-        conn.execute(
-            "create table if not exists variables (
-                id              integer primary key,
-                value           integer not null
-            )",
-            NO_PARAMS,
-        )
-        .expect("Create variables table");
-
-        DataBase {
-            conn: conn
-        }
-    }
-
-    pub fn get_is_season(&self) -> ServerResult<bool>
-    {
-        GET_OR_CREATE_DB_VAR!(&self.conn, IS_SEASON_ID, 1)
-            .map(|num| num == 1)
-    }
-
-    pub fn set_is_season(&self, val: bool) -> ServerResult<()>
-    {
-        if val == true
-        {
-            self.start_new_season(true)?; // CAREFUL
-        }
-
-        let val = if val == true { 1 } else { 0 };
-        self.conn.execute("update variables set value = (?1) where id = (?2)", params![
-            val,
-            IS_SEASON_ID
-        ])?;
-        Ok(())
+        DataBase::init(file)
     }
 
     pub fn create_superuser(&self, name: String) -> ServerResult<()>
@@ -1282,6 +1160,8 @@ mod test
 {
     use rusqlite::NO_PARAMS;
     use super::*;
+    use crate::user::{USER_ROLE_INACTIVE, USER_ROLE_REGULAR, USER_ROLE_SOFT_INACTIVE, USER_ROLE_SUPERUSER};
+    use crate::server_season::N_SEASON_ID;
 
 
     fn respond_to_match(s: &DataBase, name: &str, id: i64)
